requirements.txt
chromadb
langchain
langchain_community
langchain_ollama
ollama
marker-pdf
nltk
numpy
Pillow
pymupdf4llm
python-dotenv
rank_bm25
reportlab
Requests
scikit_learn
sentence_transformers
streamlit
